Rewrite inter protocol code
Write the protocol code as its own module with proper encoding and decoding.  It
is written in a rather "Rusty" way which probably results in quite a bit of
extra copying.  Consider moving the event queue out into the outer structure so
that it doesn't have to be copied around until the end.