Use `Vec` for EventVec
The protocol code does quite a bit of moving of the EventVec, and most of the
time, this is either empty, or only contains a small number of events.  Change
this to use an allocated `Vec` instead.  This does require an allocator, but
this should be both better use of memory as well as better use of cpu time.