Implement a prefix-search on the dictionary
In order to do steno lookups of arbitrarily typed user-text, we need to be able
to do lookups that only need to match a prefix of what was entered.  This is a
little complicated, and I've added a small `prefix-search` crate that shows the
algorithm using strings.  The prefix search multiples the search time by the
number of strokes that need to be searched for.  This will be fairly small most
of the time.