main: Factor out main function
Separate main with the main loop from the part that initializes things.  Due to
the way gpios are moved out of the pins structure, these have to actually be
used within main itself, or passed entirely to an initializer.

To start this process, remove the loop part of main into its own function, after
passing everything into it.  This requires a fairly unpleasant signature on this
function, so is probably not the best way to go about this, but there probably
isn't any way around this, short of trying to abstract these things more.  If we
abstracted out the initialization, that routine would end up with the messy
signature, along with the types that stored the result.