layout: Create a qwerty mode
The qwerty mode implements a more traditional type of keyboard layout, albeit
with more complexity of layers and key combinations.  This implements a single
key 'q', to make sure we can switch modes and that the new way of sending
keysets works.

This currently uses a Vec to hand around the keyset, and the current keyset is
managed as a BTreeSet.