qwerty: Implement combo keys
Handle combo keys.  This is done in a way that hopefully will prevent the weird
stuck keys that I occasionally will see with the qmk version.  The combinatons
are all done at the done at the scancode level.

There is some fairly complicated tracking here to make sure that we are ready to
process the release of the combos in the same layer the down even was processed in.