qwerty: Add some additional keys to num layer
Add in some of a-f, x, -, and '.' to the number layer.  Unfortunately,
the letter 'e' is already '3' so that doesn't make sense.  To make this work,
put the 'e' where the letter 'w' is.  It might take a little getting used to,
but at least it will be available without releasing the number lock.