qwerty: Handle release after layer change better
Before this change, if the layer change happens before a key-release, that
release won't be handled in the right layer, and might get stuck.  An easy way
to reproduce this is to hold press NUM-1-releaseNUM-release1, all fairly
quickly (before the host has a chance to repeat any keys).  The 1 will then
begin repeating because it hasn't been sent as a release.

We fix this by storing additional information on each key that is down.  First,
we store them by their scancode instead of resulting key.  With each entry, we
store both the particular mapping for that code, as well as the layer we got
that mapping from.

It isn't actually necessary to store the layer, as we do have the mapping, and
could just use it instead of looking it up.