qwerty: Simplify release handling
We don't actually need to store both the mapping and a reference to the layer
where we got the mapping from, since all we're going to do with that layer is
look up the mapping.  Instead, just store the mapping, and only do the lookup if
we didn't get a mapping.