layout: Indicate in-process mode select
Since mode select is a held mode, it would be helpful to indicate the possible
new mode through the indicator LED.  Implement this with a new even that
indicates we are in the process of selecting a new mode.  It flashes the LED
with the new mode.  This is useful when doing mode select and selecting a
specific mode.