Add a debugging support
The 'stall' routine can be called with a non-zero value to allow the system to
stall early, and then be continued to debugged.  This works around an issue with
the pico probe not being able to do early debugging.

To debug, change the value in GATE to 0.  Then add a call to `stall(1)`
somewhere.  Run the code and then attach the debugger, it should be stopped in
the `stall` function.  After setting breakpoints and such, you can use

    set GATE.v.value = 1

to allow the stall gate to pass.