proto: Give priorities to the IRQ handlers
Give the IRQ handlers both higher priority than the main tasks.  Assign USB 3
and UART 4.  It is hard to get an exact measurement of what the deadlines are
for these.  The UART needs to be handled before overflow, which assuming
32-bytes of FIFO at around 400kbps, is about 800us.

USB deadlines are probably longer, other than the initial message.

Realistically neither of these handlers takes very long, and everything should
be well within the deadlines.