usb: Send release after everthing typed
Although it seemed kind of like an optimization to avoid sending a key release
when we are typing ordinary text.  However, to avoid duplicates, we need to
detect this, and send a release in between.  It is easier to just always send a
release after each character we type.  The character speed to the host should be
sufficient to handle this.