steno: Start of RamDict support
This begins the transition from the full-sequence lookup, to a stepped-lookup
which will work like an NFA.  Each added stroke narrows the range of dictionary
entries considered.

To make this work, the RamDict can't just be a plain dictionary, and we instead
build a packed dictionary similar to how the mapped dictionary works.