steno: Send raw steno for untranslates
Although these will usually get replaced with a definition, now change so that
we are always sending something every time a stroke is received.  A future
change may allow these inserts and deletes to be elided (perhaps with a timer),
but at this point, just write them and delete them.