keyboard: qwerty: Add Fn layer
Add a layer map for the left two thumb keys that makes the top row into Function
keys.  The are numbered to keep in line with the numbers, so F12 is over the
leftmost key, 0 is F10, and the '-' key is F11.  This mirrors how this was done
in the QMK firmware.