keyboard: Shift steno up a row
Instead of steno keys being effectively shifted down on the keyboard from
qwerty, shift them up a row. This puts the number keys below them, which is
mostly useless, but they aren't really used with Lapwing anyway.  Giving this a
try to see how hard it is to get used to.