proto: Add USB/ACM support for Gemini
Change the "raw" steno mode from one that types the raw strokes directly, to use
the Gemini protocol over ACM.  This mode works a lot better with Plover than
NKRO which needs a stroke to enter and leave the mode.  We can support raw mode
with a dictionary solo in plover.