taipo: Properly support held keys
Instead of immediately sending a keyrelease after each key, try to do it when
the key itself is release.  We have to send the release early, though, before
another key is sent, because there can be duplicates with Taipo (press the same
key on both sides).

This is almost the right thing. In the sequence left-press, right-press,
left-release. The right key will be released instead.  I'll leave this for now,
as it will be a bit more complex to track which release goes with which
key (since the same key could be pressed on both sides).  If this becomes too
irritating, it could be fixed.