tiapo: Add workaround for typing.com
The typing.com typing tutorial doesn't accept shifted characters when the shift
key is pressed at the same time as the key being pressed.  Implement a simple
workaround by sending the modifier before the key itself.  Only do this for
shift keys, as these are the only ones that seem to matter here.