usb: Add the ability to "stall" when sending keypresses
typing.com is more irritating than just wanting the shift key to be separate, it
seems to actually require a bit of a delay. Add a "stall" notion to the key
event system to support this, and insert a 50ms delay between the shift and the
character.  This should probably be made conditional so that normal use of the
keyboard doesn't result in these stalls.