zbbq: Enable very basic HID keyboard supported
Interface to the Zephyr USB stack primarily through C code, and provide two
simple calls to this C code from Rust. One queries if all reports have been
sent, meaning it is safe to send the next report, and the other sends a report.

This simplifies the actual reports, not inserting any modifier keystrokes, but
just setting the modifier bits themselves.  This does seem to work in Windows.