zbbq: Support reverse scanned matrices
It is possible to add twice as many keys to a matrix by placing two keys on each
intersection, each with their diodes in opposite directions.

This is indicated with a dt property.

To make this work, we need to scan differently, not using low-driven signals,
but pull-downs, which requires a small delay after releasing each driven line.
Then the scan proceeds twice, once driving colums and reading rows, and then
again driving the rows and reading the columns. As long as the input threshold
has a Schmitt trigger, and the 3-diode voltage drop threshold is insufficient to
reach the threshold, but one drop is, this will then work.  It seems to work
fairly well on the RP2040.