zbbq: Change the event queue to use Mutex/Condvar
Use the new Mutex and Condvar types.  This is a little tricky, as the
initialization is no longer const.  For now, use unsafe access to a `static
mut`.  Eventually, we'll probably want to implement Once and friends (maybe even
`lazy_static`).

For now, there is an accessor function to access it.  This ultimately needs this
static, as it needs to be accessed by callbacks (which don't always take an
argument).