zbbq: Better Zephyr/Rust integration
Depends on a CONFIG_RUST feature within Zephyr. Include support for capturing
the Kconfig settings from the Zephyr build and making them available in the
application.  This will eventually be used within some of the zephyr specific
crates, and not really needed by applications.

This also generates a `build/rust/sample-cargo-config.toml` file that can be
copied or symlinked into the source directories `.cargo/config.toml` file to
allow the Rust code to be build with just plain cargo commands. This is
important to have rust-analyzer and various IDE tools work while editing this
code.