zbbq: Attempt to set features based on shield
Use the shield to determine features for 2 vs 3 row keyboards.  This doesn't
work with the .cargo/config.toml sample file, so IDE builds won't have the right
features set (they will just be from the gui).  It is probably best to figure
out how to set those in the IDE.