zbbq: Initial uart mode
This supports inter-board communication using a uart.  There is some debugging
cruft possibly left behind, finding what ultimately was a bug in the Zephyr
rp2040 uart driver, which was not enabling the FIFO.

This now uses the fifo to buffer data between 1ms polls, and otherwise keeps the
same code and logic as the 'proto' code. This has only been tested on the active
side (with the USB connector), and a subsequent patch will make that code work.