zbbq: Make led thread
The rp2040 ws2812 pio driver is blocking and polled.  As the API is defined as
blocking, we need to create a thread to do the actual updates to the LED. As
trhe driver is currently spinning on the status, we need to make this a
low-priority thread.

Implement this low priority thread, with a static Mutex used to communicate with
the main thread to indicate the state the LEDs should be set to.  This reduces
about 460us off the time taken in the main loop.  The LED updates are made
infrequent, as we aren't doing any dynamic-changing PWM at higher rates.