keyboard: taipo: add function keys
Instead of the Taipo way for the function keys, with a prefix, and then the
numbers, just use the numbers with both thumb keys for the function keys.  We
need two additional for F11 and F12, and will use 'v' and 'w' respectively.
These are slightly mnemonic (a bit of a stretch), but importantly, don't overlap
with any of the existing keys.