zbbq: Make LED thread high priority
Although the LEDs on the RP2040 use the pio to do the actual write, the data
still has to be fed byte-by-byte, and if there is a stall, will result in
corrupting what is written to the LEDs.

Fix this by making the LED thread a high priority non-preemptible thread.  If
the driver later gains DMA capabilities, this would no longer be necessary.

This can cause a stall of up to about 300us in the main scan loop.  Interrupts
can still corrupt the LED values written, but this change makes them _usually_
write correctly.