zbbq: use `addr_of_mut`
Rust 1.77 now warns about taking the address of a mutable static.  It seems that
the intermediate reference before the raw pointer is taken is considered unsafe,
and this will be deprecated in the future.  Fortunately, since 1.51 there has
been a macro `addr_of_mut` that can make a raw pointer from a static without the
problem.

Convert all warned references to statics to this macro to eliminate the warnings.