jolt: Use send_boxed from irq
The irq should not be calling malloc (which Box does).  Instead, we will
have the main thread allocate a box for it, and place it in a static
RefCell protected by a critical section.

This then uses the new `send_boxed()` method on `Sender`, which allows
the allocation to take place at an earlier time than the actual send.