bbq-steno: joiner: Fix state carryover
The previous state needs to go back as many strokes as our length to
find the proper state.  This was causing things like removal due to
replacements to have the state affect things, regardless.

In addition, carry over the space state, so that translations that don't
isnert any text will still have the right state in place.