jolt: update some LED colors
The LED on the jolt1 is fairly dim, so brighten up the colors a bit.
Ideally, each keyboard or LED would have some factors to help get the
LED values. For example the ws2812 modules tend to be very bright
compared to a simple PWM-based led that is on the Tiny 2040.