jolt: leds: Generalize, with PWM first
Generalize the LED manager to work generically, instead of with a
specific driver.  Start with the PWM driver (WS2812 will be in a
subsequent patch).

This does change the LED state data to a Vec, which will result in
allocations.  The sizes should all be the same, so there shouldn't be
fragmentation issues.