bbq-keyboard: Fix key rollover to modifier
Previously, rolling into a modifier key quickly would result in the
modifier being spit out as separate keys instead of being seen as a
modifier.  Instead of just sending the keys when a new key isn't a
combo, use the opportunity to send the previous one, and then treat the
new key as a potential match of a combo.