jolt: Add thread utilization stats
This collects stats on how long each task takes to do its job.  This is
real time, so lower priority tasks will be obscured by higher priority
tasks running, but this is basically what we need to know.  If the main
loop takes more than 1ms, ticks will be missed.