jolt: matrix: spread out matrix scanning
To reduce CPU load, spread out the matrix scanning to one column per 1ms
tick.  This has the advantage of using less CPU.  But it also reduces
the debouncing to only 4 samples.  Try this for a while, and if I see
any bouncing (or delayed keys), revert this, as the CPU burden isn't
actually all that great.