Move Dict casting down
Instead of having Memdict::from_raw_ptr return concrete types and each
use cast those in to the dyn Trait boxes, do this in this routine.

This cleans up the code a little bit, but more importantly, allows a
future change to encode internal dictionaries which will be represented
differrently.