bbq-keyboard: Idempotent inter protocol
Replace the array of key events with a bit map indicating all of the
keys that have been pressed.  This handles dropped packets much better
than the stuck or missed keys that is seen with the previous protocol.

There is no versioning information, so it is important that both sides
are upgrade together.  Otherwise spurious and incorrect events will
likely be seen.