bbq-keyboard: serialze: Fix bound error when corrupt packet
Because we run the decode state machine before decoding the CRC, it is
important to not panic on corrupt packets.  In this case, discard the
error we would get trying to push to an ArrayVec past the end.  In all
likelihood, this packet will have a bad CRC, but we need to not panic.