jolt: Send HID plover reports
These keys appear correct in the Plover HID Inspector (despite the bits
being bit endian), but I don't have the Plover HID pluging
(plover-machine-hid) itself working, so I can't actually test this.

It turns out that the Plover install on MacOS is an x86_64, and when
homebrew tries to install the hidapi library, it only installs the
native arm64, so Plover can't see it.  At this point, it looks like I'm
probably best focusing on improving on-board steno.