jolt: Add start of keyminder protocol handler
This starts the handling of the keyminder protocol.  This was intended
to be used with hidapi on the host.  It turns out that this is fairly
useless, as hidapi assumes that out reports are just status updates, and
can be freely discarded (without even giving status) if the device isn't
ready to receive the report.

So, this won't actually work, sadly, and we'll probably need to use
libusb to talk to a bulk endpoint pair where we can implement our own
protocol.