jolt: Convert the minder to use the second ACM
Use the second ACM port for the inder protocol.  This is a bit awkward
on the host side, as the is no easy way to correlate which acm port goes
with which HID device.  Realistically, we'll probably want something to
do an LED blink based on the host program to help identify the right
keyboard.

For now, this removes the use of the other HID endpoing, but doesn't
remove it.

This will probably eventually be better as a USB Bulk endpoint pair, but
there is a bunch of Zephyr work needed to support that.