jolt: Change uart write to async
Change minder to use the new async write api.  Instead of writing
directly, write to a buffer, and then enqueue that as a single write.
We just drop the buffers when writes complete.  We avoid reading from
the log if the write buffer is full, so that the logging system will
probably record that messages have been dropped.