jolt: Read keyboard HID out report
As we have a descriptor for this report, the host will be sending the
report for LED status changes (notably, caps lock).  If we don't read
the report, the host OS might not accept any more key reports from us.

On MacOS, this seems to cause the keyboard to stop working when it is
plugged directly into the computer.  With a hub, it seems the hub
accepts the report, to forward itself.  This might be a feature of USB
3.

Fix this by actually reading the report.  It only logs a message and
doesn't do anything with it (we could consider an LED update).