bbq-keyboard: layout: bypass Plover mode
Big move.  Remove the plover mode from the supported modes.  The mode
select on 2 row keyboards is between taipo and steno.  On the three row,
it cycles between qwerty, steno, and taipo.

Plover mode can still be selected with direct keys. But, as I don't seem
to be using it any more, it is less useful to have in the normal
rotation.