jolt: Move to new inter protocol
Use the "ser2" protocol instead of the old "serialize" protocol.  This
is built around CBOR (using mini-cbor) instead of the ad-hoc protocol
from before. The sequence number has been removed, as it didn't actually
do anything.

The new protocol uses a struct or the packets with optional fields
rather than an enum.  This will make it easier, in the future, to stop
sending the packets when we know both sides have a coherent view of
state.  This is not yet implemented.