jolt: Change to bounded channels
Use bounded channels.  This simplifies the timer irq handler, as we can
now just directly send a message from the handler.  Change the old
handoff messaging to just use a simple semaphore (keeps the ticks from
building up, as ticks will be skipped if the previous one is still
processing.