bbq-keyboard: layout: Add "Insert" key
Add bindings to both Qwerty and Taipo for the Insert key.  For Qwerty,
use the 'n' key in the NAV layer.  For Taipo it is tab (index, middle
ring on top row) with the "far" thumb key (the one closes to the center
of the keyboard, furthest away from the home row).

The key is actually kind of pointless, as MacOS doesn't support it, and
it ends up being translated to "Help".  As such, nothing will actually
use it, but "Help" can be bound, giving rise to an additional key.