jolt: First pass at moving to async
Create a worker thread to run these tasks.  Move mainloop, the matrix
scan, and the inter handling code to their own async tasks.

By adjusting the timing of the inter task to only run every 5ms, we now
finally are under 100% CPU and normal matrix scans aren't skipped.  The
slow inter scan still causes them to be dropped, so it might make sense
to move this to a lower-priority worker (maybe with leds or something,
who knows).

But, the keyboard seems to work with this.  Next is to move the other
periodic scanning tasks into their own async task, and basically
eliminate the main loop altogether.  There is still the  'main' event
queue that needs to be redirected, but changing that will require
incompatible changes to bbq-keyboard.