jolt: Simplify USB HID event creation
Simplify the flow of events from decoded keys to encoding of events.
Before this change, when layout or steno wishes to send key events, the
following happens:

- layout/steno sends a KeyAction to the main event queue
- The main loop reads this event, and adds it to a VecDeq for "keys".
- At the end of the main loop, it calls a `usb_hid_push` to push these
  to the USB layer

This hid push only happens on the 1ms tick, but the Usb already handles
the events needing to be queued up, so this extra VecDec doesn't help
anything, and just adds more unnecessary processing.

Fix this by making `usb_hid_push` just take a single KeyAction, and
calling it directly as keys come in.

This is just the first step toward making layout and steno themselves
actually queue the events directly with the Usb HID code.