bbq-keyboard/jolt: Change layout's tick to take a tick count
To allow this to be called less frequently, accept a "ticks" argument
which is the number of ms that has passed.  This allows us to ping this
less frequently.  Given that all of the current timeouts are 50ms, this
shouldn't have any effect, other than making the 50ms from +/- 1ms to
+/- whatever the tick intervals is.

Pass the argument from jolt, and start with a 5ms timeout, which seems
to eliminate the overflows we've seen.