jolt: Move led manager into Dispatch
Moves it in, but for now, just keep it public.  This also moves
"rawmode" and "current_mode" into Dispatch, for now wrapping with
SpinMutex.  Adds locks around them in the current main loop.

This also comments out the remaining Gemini code, which isn't currently
active.  This can be added back in separately entirely within Dispatch