jolt/bbq-keyboard: Increase async of layout manager
Make the layout manager events not use the event queue, but instead as
direct calls to async functions.  This changes the api to the manager to
be async, and use a specific trait to provide the methods.

Removes these events from the Event list, which ensures that nothing
else will use the event queue for this.

Currently, doing this is causing "Periodic 1m overflow" messages from
the dispatch event loop, so some analysis might be needed as to why.
This is likely coming from layout's tick sending USB events, which
shouldn't might be rescheduling.