bbq-steno: Properly clean steno history
The pieces were in place to prune the steno history, but it didn't seem
to actually get implemented.  Fix this by ensuring that we don't store
too much history.  This should result in fairly constant memory use
after the limit has been reached.  It is currently 50 strokes, which
realistically, is almost absurdly large.