jolt-embassy-rp: Enable systemview within 'log'
Systemview takes over the RTT port, and it is easiest to just enable
this with the 'log' framework.  Currently, it is always enabled,
although this could also be done with another crate.

With this, SystemView is able to see task scheduling.  With some patches
to embassy-rp, it will also show time in ISR handlers.