jolt-embassy-rp: Implement inter protocol
This adds an I2C based inter-board protocol.  The I2C bus is used, with
the active side as the controller, to communicate the state of keys from
the passive side.  The TX line from the passive side's uart channel is
used as an IRQ (active high) so that we don't need to constantly poll.