minder: Add a simple cobs encoder/decoder
COBS is a way to handle packet framing and stuffing with fixed overhead.
It work as long as the total packet size is less than 256 bytes (it can
be extended for more, but loses the fixed over).

For each transmitted packet, two additional bytes are needed for the
framing, and stuffing.