jolt-embassy-rp: Add support for some jolt2 boards
This includes the jolt2dir Left side, and the jolt2 (Pimoroni version)
right side.  The main difference is that these boards use a serial
protocol, which is a fresh implementation here (distinct from the Zephyr
protocol). This protocol will stop transmitting shortly after each
update, and the send and receive sides are uncorrelated.