jolt-embassy-rp: Clean up inter-board variants
There are now four variants of how we communicate with another keyboard
half, the cross product I2C, Uart, and active-passive.  The two
implementations are quite different.

Instead of four Options and some potentially strange code to decode
them, make this an enum to more clearly map to only one (and exactly
one) being available at any time.

We could also add a "None" element, when adding support for the proto4,
which is a two row keyboard, with only a single CPU.