bbq-keyboard: Implement Taipo/Steno mixed
Add support to a Taipo layer shift while in Steno mode.  We use the
upper '+' and '^' keys for this.  By either pressing them with a stroke,
or holding one down and stroking, the keys will be interpreted in the
Taipo keyboard layout.

This makes a for a quick way to send raw strokes.  The full taipo
engine is used for this, so modifiers and such also work.

Full mode switching is done via the existing mode switch, although it
may make sense to use a press and release of the taipo shift to switch
between steno and taipo.