jolt-embassy-rp: Back off the inter hello
When the protocol is failing (such as the other side being disconnected,
or debugged), back off of the initial hello packet, with a max of 10
minutes.  This avoids flooding the log with messages about this failed
exchange, without cauing the messages to be missed entirely.

Note this only happens after for the initial hello. We assume that when
the interrupt is asserted, the device will be present to query the data
from it.